requirements.txt
nemoguardrails
openai
inspeqai
python-dotenv
fastapi==0.111.0
mangum==0.17.0
numpy==1.24.4
pydantic==2.7.4
python-dotenv==1.0.1
python_json_logger==2.0.4
starlette==0.37.2
uvicorn==0.30.1
protobuf==4.25.1
absl-py==2.1.0
colored==2.2.4

gensim